Move display models into their own module
