+++ archive/Makefile
@@ -5,7 +5,7 @@ STACK_ROOT 	= archive
 
 SBT_APPS    = archivist progress_async progress_http notifier registrar_async registrar_http
 SBT_DOCKER_LIBRARIES = archive_common progress_common
-SBT_NO_DOCKER_LIBRARIES = registrar_common
+SBT_NO_DOCKER_LIBRARIES = registrar_common archive_display
 
 ECS_TASKS   = bagger archive_api callback_stub_server
 LAMBDAS     = migrator
