+++ archive/display/src/main/scala/uk/ac/wellcome/platform/archive/display/DisplayProvider.scala
@@ -1,10 +1,12 @@
-package uk.ac.wellcome.platform.archive.common.models
+package uk.ac.wellcome.platform.archive.display
 
 import io.circe.generic.extras.JsonKey
 import uk.ac.wellcome.platform.archive.common.progress.models.StorageProvider
 
 case class DisplayProvider(id: String,
-                           @JsonKey("type") ontologyType: String = "Provider")
+                           @JsonKey("type") ontologyType: String = "Provider") {
+  def toStorageProvider: StorageProvider = StorageProvider(id)
+}
 object DisplayProvider {
   def apply(provider: StorageProvider): DisplayProvider =
     DisplayProvider(id = provider.id)
