+++ archive/display/src/test/scala/uk/ac/wellcome/platform/archive/display/DisplayIngestTest.scala
@@ -1,4 +1,4 @@
-package uk.ac.wellcome.platform.archive.common.models
+package uk.ac.wellcome.platform.archive.display
 
 import java.net.{URI, URL}
 import java.time.Instant
@@ -7,10 +7,11 @@ import java.util.UUID
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.platform.archive.common.fixtures.RandomThings
+import uk.ac.wellcome.platform.archive.common.progress.fixtures.TimeTestFixture
 import uk.ac.wellcome.platform.archive.common.progress.models._
 import uk.ac.wellcome.storage.ObjectLocation
 
-class DisplayIngestTest extends FunSpec with Matchers with RandomThings {
+class DisplayIngestTest extends FunSpec with Matchers with RandomThings with TimeTestFixture{
 
   private val id = UUID.randomUUID()
   private val callbackUrl = "http://www.example.com/callback"
@@ -56,5 +57,31 @@ class DisplayIngestTest extends FunSpec with Matchers with RandomThings {
     ingest.lastModifiedDate shouldBe modifiedDate
     ingest.events shouldBe List(
       DisplayProgressEvent(eventDescription, eventDate))
+  }
+
+  it("transforms itself into a progress") {
+    val displayProvider = DisplayProvider("s3", "Amazon s3")
+    val bucket = "ingest-bucket"
+    val path = "bag.zip"
+    val progressCreateRequest = RequestDisplayIngest(
+      DisplayLocation(displayProvider, bucket, path),
+      Some(
+        DisplayCallback("http://www.wellcomecollection.org/callback/ok", None)),
+      DisplayIngestType("create"),
+      DisplayStorageSpace("space-id")
+    )
+
+    val progress = progressCreateRequest.toProgress
+
+    progress.id shouldBe a[UUID]
+    progress.sourceLocation shouldBe StorageLocation(
+      StorageProvider(displayProvider.id),
+      ObjectLocation(bucket, path))
+    progress.callback shouldBe Some(
+      Callback(URI.create(progressCreateRequest.callback.get.url)))
+    progress.status shouldBe Progress.Initialised
+    assertRecent(progress.createdDate)
+    assertRecent(progress.lastModifiedDate)
+    progress.events shouldBe List.empty
   }
 }
