+++ build.sbt
@@ -185,11 +185,15 @@ lazy val archive_common = doServiceSetup(project, "archive/common")
   .dependsOn(config_storage % "compile->compile;test->test")
   .settings(libraryDependencies ++= Dependencies.archiveCommonDependencies)
 
+lazy val archive_display = doServiceSetup(project, "archive/display")
+  .dependsOn(archive_common % "compile->compile;test->test")
+
 lazy val archivist = doServiceSetup(project, "archive/archivist")
   .dependsOn(archive_common % "compile->compile;test->test")
 
 lazy val notifier = doServiceSetup(project, "archive/notifier")
   .dependsOn(archive_common % "compile->compile;test->test")
+  .dependsOn(archive_display % "compile->compile;test->test")
   .settings(libraryDependencies ++= Dependencies.wiremockDependencies)
 
 lazy val registrar_common = doServiceSetup(project, "archive/registrar_common")
@@ -200,6 +204,7 @@ lazy val registrar_async = doServiceSetup(project, "archive/registrar_async")
 
 lazy val registrar_http = doServiceSetup(project, "archive/registrar_http")
   .dependsOn(registrar_common % "compile->compile;test->test")
+  .dependsOn(archive_display % "compile->compile;test->test")
 
 lazy val progress_common = doServiceSetup(project, "archive/progress_common")
   .dependsOn(archive_common % "compile->compile;test->test")
@@ -210,7 +215,7 @@ lazy val progress_async = doServiceSetup(project, "archive/progress_async")
 
 lazy val progress_http = doServiceSetup(project, "archive/progress_http")
   .dependsOn(progress_common % "compile->compile;test->test")
-  .settings(libraryDependencies ++= Dependencies.wiremockDependencies)
+  .dependsOn(archive_display % "compile->compile;test->test")
 
 lazy val root = (project in file("."))
   .aggregate(
@@ -254,6 +259,7 @@ lazy val root = (project in file("."))
     snapshot_generator,
 
     archive_common,
+    archive_display,
     archivist,
     notifier,
     registrar_async,
